Moved maxTopics one layer down in the xml, so it is under <topic><max-topics> and not <topic><topics><max-topics>. Added a placeholder for the nested config entity, if it is needed in the future
