+++ broker/src/main/java/com/hivemq/configuration/service/TopicConfigurationService.java
@@ -1,12 +1,5 @@
 package com.hivemq.configuration.service;
 
-import com.hivemq.annotations.ReadOnly;
-import com.hivemq.configuration.service.entity.*;
-import com.hivemq.configuration.service.exception.ConfigurationValidationException;
-import com.hivemq.mqtt.message.QoS;
-
-import java.util.List;
-
 /**
  * The service which allows to inspect Topic configuration at runtime.
  *
